Finish setting up checklist for shopping list items
